Enable configuration of NR UE through a YAML file
This commit adds libyaml-cpp-dev to required docker images and build commands. There is one workaround
in cross-compilation for apt behavior for installing libyaml-cpp-dev. Apt will uninistall previous
versions of the package whenever another one is installed. This means that installing the arm version
breaks amd version and vice versa. This makes it impossible to have both libs installed at the same time
which is required by the cross-compilation build command (it calls cmake twice, once with native and once with
target architecture)