Refactor UE CSI procedures
CSI IM power estimation is split into two functions. One sums the power
of all REs in each symbol and is called each symbol and the other is
called in the last symbol to compute the average power.

CSI RS procedures is also split into two parts. The first part computes
CSI RS estimates and the other part does the measurements and reports to
MAC.